created an endpoint that return both available and unavailable spaces
