+++ utils/BookingStatus.js
@@ -2,5 +2,5 @@ export const BOOKING_STATUS = {
     PENDING: "PENDING",
     FULFILLED: "FULFILLED",
     CANCELLED: "CANCELLED",
-    INUSE: "IN-USE"
+    IN_USE: "IN_USE"
 }
